Add automatic window sizing
Add a mechanism for automatically adjusting the size of the rendering window to
match the bounding box of the fractal exactly. This removes the need to define a
`min_window` attribute for a fractal definition.

This commit also updates the README and example fractals accordingly, and
removes the issue that with `quiet` mode enabled the helper functions in
`make_raster` would not be properly called.